codex: keep CLI cleanup guard local
